+++ plugin/src/main/java/com/stratio/cassandra/lucene/common/GeoDistanceUnit.java
@@ -15,7 +15,7 @@
  */
 package com.stratio.cassandra.lucene.common;
 
-import org.codehaus.jackson.annotate.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonCreator;
 
 /**
  * Enum representing a spatial distance unit.