+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/builder/GeoShapeMapperBuilder.java
@@ -15,11 +15,11 @@
  */
 package com.stratio.cassandra.lucene.schema.mapping.builder;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.stratio.cassandra.lucene.common.GeoTransformation;
 import com.stratio.cassandra.lucene.common.JTSNotFoundException;
 import com.stratio.cassandra.lucene.schema.mapping.GeoPointMapper;
 import com.stratio.cassandra.lucene.schema.mapping.GeoShapeMapper;
-import org.codehaus.jackson.annotate.JsonProperty;
 
 import java.util.Arrays;
 import java.util.List;