+++ plugin/src/test/java/com/stratio/cassandra/lucene/common/GeoTransformationBufferTest.java
@@ -16,7 +16,6 @@
 package com.stratio.cassandra.lucene.common;
 
 import com.spatial4j.core.shape.jts.JtsGeometry;
-import com.stratio.cassandra.lucene.util.JsonSerializer;
 import com.vividsolutions.jts.geom.Geometry;
 import org.junit.Test;
 
@@ -25,8 +24,8 @@ import java.io.IOException;
 import static com.stratio.cassandra.lucene.common.GeoDistanceUnit.KILOMETRES;
 import static com.stratio.cassandra.lucene.common.GeoDistanceUnit.METRES;
 import static com.stratio.cassandra.lucene.common.GeoTransformation.Buffer;
-import static com.stratio.cassandra.lucene.util.GeospatialUtilsJTS.CONTEXT;
-import static com.stratio.cassandra.lucene.util.GeospatialUtilsJTS.geometry;
+import static com.stratio.cassandra.lucene.common.GeospatialUtilsJTS.CONTEXT;
+import static com.stratio.cassandra.lucene.common.GeospatialUtilsJTS.geometry;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;